IdMgr Define

+++ oneflow/common/range.h
@@ -11,19 +11,19 @@ class Range final {
   Range() : Range(0, 0) {}
   ~Range() = default;
 
-  Range(int32_t begin, int32_t end) : begin_(begin), end_(end) {}
+  Range(int64_t begin, int64_t end) : begin_(begin), end_(end) {}
 
-  int32_t begin() const { return begin_; }
-  int32_t end() const { return end_; }
+  int64_t begin() const { return begin_; }
+  int64_t end() const { return end_; }
   
-  int32_t& mut_begin() { return begin_; }
-  int32_t& mut_end() { return end_; }
+  int64_t& mut_begin() { return begin_; }
+  int64_t& mut_end() { return end_; }
 
-  int32_t size() const { return end_ - begin_; }
+  int64_t size() const { return end_ - begin_; }
 
  private:
-  int32_t begin_;
-  int32_t end_;
+  int64_t begin_;
+  int64_t end_;
 };
 
 } // namespace oneflow
