+++ oneflow/common/shape.h
@@ -20,15 +20,15 @@ class Shape final {
   // Getters and Setters
   const std::vector<int64_t>& shape_vec() const { return shape_vec_; }
   int64_t elem_cnt() const { return elem_cnt_; }
-  int64_t At(int32_t index) const;
-  void Set(int32_t index, int64_t val);
-  int32_t NumAxes() const { return shape_vec_.size(); }
-  int64_t Count(int32_t start_axis, int32_t end_axis) const;
-  int64_t Count(int32_t start_axis) const;
+  int64_t At(int64_t index) const;
+  void Set(int64_t index, int64_t val);
+  int64_t NumAxes() const { return shape_vec_.size(); }
+  int64_t Count(int64_t start_axis, int64_t end_axis) const;
+  int64_t Count(int64_t start_axis) const;
 
  private:
   void UpdateElemCnt();
-  int64_t CanonicalAxisIndex(int32_t axis_index) const;
+  int64_t CanonicalAxisIndex(int64_t axis_index) const;
 
   std::vector<int64_t> shape_vec_;
   int64_t elem_cnt_;
@@ -44,16 +44,16 @@ inline bool Shape::operator == (const Shape& rhs) const {
   return shape_vec_ == rhs.shape_vec_;
 }
 
-inline int64_t Shape::At(int32_t index) const {
+inline int64_t Shape::At(int64_t index) const {
   return shape_vec_[CanonicalAxisIndex(index)];
 }
 
-inline void Shape::Set(int32_t index, int64_t val) {
+inline void Shape::Set(int64_t index, int64_t val) {
   shape_vec_[CanonicalAxisIndex(index)] = val;
   UpdateElemCnt();
 }
 
-inline int64_t Shape::Count(int32_t start_axis) const {
+inline int64_t Shape::Count(int64_t start_axis) const {
   return Count(start_axis, NumAxes());
 }
