+++ oneflow/operator/clone_op.cpp
@@ -10,7 +10,7 @@ void CloneOp::Init(const OperatorConf& op_conf) {
   mut_pb_op_conf().reset(cnf);
 
   EnrollInputBn("in");
-  for (int32_t i = 0; i < cnf->out_num(); ++i) {
+  for (int64_t i = 0; i < cnf->out_num(); ++i) {
     EnrollOutputBn("out_" + std::to_string(i));
   }
 }
